requirements.txt
anthropic>=0.40.0
slack-bolt>=1.20.0
flask>=3.0.0
gspread>=6.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
thefuzz[speedup]>=0.22.0
pytest>=8.0.0
